backend/requirements.txt
fastapi
uvicorn
openai
psycopg2
python-dotenv